Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -15,7 +15,7 @@ int main() {
             printf("%d",a*b);       
             break;
         case '/':
-            printf("%d",int(a/b));
+            printf("%d",a/b);
             break;    
         default:
             printf("Error");
